requirements.txt
Django==5.1.3
django-rest-framework==0.1.0
drf-spectacular==0.27.2